+++ src/screens/HomeScreen.js
@@ -1,12 +1,20 @@
 import React, { useEffect } from 'react';
-import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
+import { View, Text, TouchableOpacity, StyleSheet, Dimensions } from 'react-native';
 import { collection, getDocs } from 'firebase/firestore';
-import { db } from './src/lib/firebase'; // adjust path if needed
+import { db } from '../lib/firebase'; // adjust path if needed
+
+import global from '../styles/global';
+import HomeScreenYard from '../assets/homeScreen/homeScreenYard.svg';
+import House from '../assets/homeScreen/houses/basicHouse.svg';
 
 export default function HomeScreen({ navigation }) {
+  const { width, height } = Dimensions.get('window');
+
   useEffect(() => {
     const testFirestore = async () => {
-      const querySnapshot = await getDocs(collection(db, 'rooms'));
+      console.log('testing firestore')
+      console.log('db is', db)
+      const querySnapshot = await getDocs(collection(db, 'user', 'VuoNhIFyleph42rgqis5','rooms'));
       querySnapshot.forEach((doc) => {
         console.log(doc.id, doc.data());
       });
@@ -15,12 +23,17 @@ export default function HomeScreen({ navigation }) {
   }, []);
 
   return (
-    <View style={styles.container}>
+    
+    <View style={global.container}>
+      <HomeScreenYard
+        width={width}
+        height={height}
+        preserveAspectRatio="xMidYMax slice" />
+      <House
+        width={300} />
       <TouchableOpacity
-        style={styles.roomBox}
         onPress={() => navigation.navigate('Room')}
       >
-        <Text style={styles.roomText}>Living Room</Text>
       </TouchableOpacity>
     </View>
   );
